refactor(test): 抽取公共函数，在测试中跳过新手引导
