+++ tests/e2e/leftRightToday.e2e.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { setUpEverything, closeEverything } from './test_utils';
+import { setUpEverything, closeEverything, skipOnboarding } from './test_utils';
 
 test('left right today', async () => {
   // 启动 Vite 开发服务器
@@ -13,20 +13,7 @@ test('left right today', async () => {
     // 该测试用例的目的是测试左侧和右侧的今天按钮功能
     await page.goto('http://localhost:5173/#/');
 
-    try {
-      const startButton = page.getByRole('button', { name: '开始使用' });
-
-      // 尝试等待按钮出现（最多等 3 秒），如果不存在则说明已完成引导
-      await startButton.waitFor({ timeout: 3000 });
-
-      if (await startButton.isVisible()) {
-        await startButton.click();
-        await page.getByRole('button', { name: '跳过引导' }).click();
-      }
-    } catch (error) {
-      // 如果“开始使用”按钮不存在，说明已完成新手引导，可跳过
-      console.log('新手引导已完成，跳过引导操作');
-    }
+    await skipOnboarding(page);
 
     await page.locator('div').filter({ hasText: /^12$/ }).first().click();
     await page.getByRole('button', { name: '' }).nth(1).click();
